pyproject.toml
[tool.poetry]
name = "yt-summarizer"
version = "0.1.0"
description = ""
authors = ["Luis Imagiire <luishenrique.imagiire@gmail.com>"]
readme = "README.md"
packages = [{include = "yt_summarizer"}]

[tool.poetry.dependencies]
python = "^3.9"
fire = "^0.5.0"
pytube = "^12.1.3"
pyyaml = "^6.0"
openai-whisper = "^20230314"
langchain = "^0.0.157"
openai = "^0.27.6"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
